+++ chaosk8s/statefulset/actions.py
@@ -1,13 +1,39 @@
 # -*- coding: utf-8 -*-
+import json
+import os.path
+
 from chaoslib.exceptions import ActivityFailed
 from chaoslib.types import Secrets
 from kubernetes import client
 from kubernetes.client.rest import ApiException
 from logzero import logger
+import yaml
 
 from chaosk8s import create_k8s_api_client
 
-__all__ = ["scale_statefulset", "remove_statefulset"]
+__all__ = ["create_statefulset", "scale_statefulset", "remove_statefulset"]
+
+
+def create_statefulset(spec_path: str, ns: str = "default",
+                       secrets: Secrets = None):
+    """
+    Create a statefulset described by the service config, which must be
+    the path to the JSON or YAML representation of the statefulset.
+    """
+    api = create_k8s_api_client(secrets)
+
+    with open(spec_path) as f:
+        p, ext = os.path.splitext(spec_path)
+        if ext == '.json':
+            statefulset = json.loads(f.read())
+        elif ext in ['.yml', '.yaml']:
+            statefulset = yaml.load(f.read())
+        else:
+            raise ActivityFailed(
+                "cannot process {path}".format(path=spec_path))
+
+    v1 = client.AppsV1Api(api)
+    v1.create_namespaced_stateful_set(ns, body=statefulset)
 
 
 def scale_statefulset(name: str, replicas: int, ns: str = "default",