+++ tests/test_statefulset.py
@@ -1,7 +1,11 @@
 # -*- coding: utf-8 -*-
 from unittest.mock import MagicMock, patch, ANY
 
-from chaosk8s.statefulset.actions import scale_statefulset, remove_statefulset
+import pytest
+from chaoslib.exceptions import ActivityFailed
+
+from chaosk8s.statefulset.actions import scale_statefulset, \
+    remove_statefulset, create_statefulset
 
 
 @patch('chaosk8s.has_local_config_file', autospec=True)
@@ -20,6 +24,7 @@ def test_scale_statefulset(cl, client, has_conf):
     assert v1.patch_namespaced_stateful_set.call_count == 1
     v1.patch_namespaced_stateful_set.assert_called_with(
         "my-statefulset", namespace="default", body=body)
+
 
 @patch('chaosk8s.has_local_config_file', autospec=True)
 @patch('chaosk8s.statefulset.actions.client', autospec=True)
@@ -64,3 +69,54 @@ def test_removing_statefulset_with_label_selector(cl, client, has_conf):
     v1.delete_namespaced_stateful_set.assert_called_with(
         "mystatefulset", "default", body=ANY)
 
+
+@patch('chaosk8s.has_local_config_file', autospec=True)
+@patch('chaosk8s.statefulset.actions.client', autospec=True)
+@patch('chaosk8s.client')
+def test_creating_statefulset_with_file_json(cl, client, has_conf):
+    has_conf.return_value = False
+
+    body = "example of body"
+
+    v1 = MagicMock()
+    client.AppsV1Api.return_value = v1
+
+    create_statefulset("tests/fixtures/statefulset/create/file.json")
+
+    assert v1.create_namespaced_stateful_set.call_count == 1
+    v1.create_namespaced_stateful_set.assert_called_with(
+        "default", body=body)
+
+
+@patch('chaosk8s.has_local_config_file', autospec=True)
+@patch('chaosk8s.statefulset.actions.client', autospec=True)
+@patch('chaosk8s.client')
+def test_creating_statefulset_with_file_yaml(cl, client, has_conf):
+    has_conf.return_value = False
+
+    body = "example of body"
+
+    v1 = MagicMock()
+    client.AppsV1Api.return_value = v1
+
+    create_statefulset("tests/fixtures/statefulset/create/file.yaml")
+
+    assert v1.create_namespaced_stateful_set.call_count == 1
+    v1.create_namespaced_stateful_set.assert_called_with(
+        "default", body=body)
+
+
+@patch('chaosk8s.has_local_config_file', autospec=True)
+@patch('chaosk8s.statefulset.actions.client', autospec=True)
+@patch('chaosk8s.client')
+def test_creating_statefulset_with_file_txt_KO(cl, client, has_conf):
+    has_conf.return_value = False
+
+    path = "tests/fixtures/statefulset/create/file.txt"
+
+    v1 = MagicMock()
+    client.AppsV1Api.return_value = v1
+
+    with pytest.raises(ActivityFailed) as excinfo:
+        create_statefulset(path)
+    assert "cannot process {path}".format(path=path) in str(excinfo.value)