+++ CHANGELOG.md
@@ -13,11 +13,13 @@
 - Added function to execute remote commands in a container
 - Remove check Kubernetes statefulset by name and label  [#70][70]
 - Added action to scale statefulsets [#73][73]
+- Added action to create a statefulset from a json/yaml file [#74][74]
 
 [53]: https://github.com/chaostoolkit/chaostoolkit-kubernetes/pull/53
 [65]: https://github.com/chaostoolkit/chaostoolkit-kubernetes/pull/65
 [70]: https://github.com/chaostoolkit/chaostoolkit-kubernetes/pull/70
 [73]: https://github.com/chaostoolkit/chaostoolkit-kubernetes/pull/73
+[74]: https://github.com/chaostoolkit/chaostoolkit-kubernetes/pull/74
 
 ## [0.21.0][] - 2019-09-02